Refactoring navigation module. Separate on bottom navigation view and bottom app bar. Add material lib.

+++ src/main/kotlin/easy/go/skel/common/resource/Dimen.kt
@@ -12,8 +12,8 @@ object Dimen {
     get() = Resources.getSystem().displayMetrics.density
 
   @JvmStatic
-  fun dpToPx(dp: Int) = (dp * density).toInt()
+  fun dpToPx(dp: Int) = dp * density
 
   @JvmStatic
-  fun pxToDp(px: Int) = (px / density).toInt()
+  fun pxToDp(px: Int) = px / density
 }
